[Warnings] Avoid warnings raised by comparison of signed/unsigned integers

+++ dune/fem/gridpart/test/test-filteredgridpart.cc
@@ -50,7 +50,7 @@ void testGridPart( const GridPartType & gridPart )
     IndexType index = indexSet.index(element);
     maxIndex = std::max( index, maxIndex);
 
-    if (index >= consecutiveIndex[0].size()) isConsecutiveIndex[0] = false;
+    if (static_cast<std::size_t>(index) >= consecutiveIndex[0].size()) isConsecutiveIndex[0] = false;
     else consecutiveIndex[0][index] = true;
     for (int c=0;c<=GridPartType::dimension;++c)
     {
@@ -59,7 +59,7 @@ void testGridPart( const GridPartType & gridPart )
       for (int i=0;i<nSubEn;++i)
       {
         IndexType index = indexSet.subIndex(element,i,c);
-        if (index >= consecutiveIndex[c].size()) isConsecutiveIndex[c] = false;
+        if (static_cast<std::size_t>(index) >= consecutiveIndex[c].size()) isConsecutiveIndex[c] = false;
         else consecutiveIndex[c][index] = true;
       }
     }
@@ -141,7 +141,7 @@ void testIntersectionIterator( const GridPartType & gridPart )
       if (intersection.neighbor())
       {
         typename GridPartType::IndexSetType::IndexType nbIndex = gridPart.indexSet().index( Dune::Fem::make_entity( intersection.outside() ) );
-        if ( nbIndex >= index.size() )
+        if ( static_cast<std::size_t>(nbIndex) >= index.size() )
         {
           std::cout << "An index on neighbor is too large" << std::endl;
           continue;
